# api-gateway/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  booking.BookingResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/booking.GeoPoint'
      provider_id:
        type: string
      scheduled_time:
        type: string
      service_id:
        type: string
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  booking.GeoPoint:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  booking.ListBookingsResponse:
    properties:
      bookings:
        items:
          $ref: '#/definitions/booking.BookingResponse'
        type: array
    type: object
...
  booking.UpdateBookingRequest:
    properties:
      id:
        type: string
      service_id:
        type: string
    type: object
...
  models.Booking:
    properties:
      location:
        $ref: '#/definitions/models.GeoPoint'
      provider_id:
        type: string
      scheduled_time:
        type: string
      service_id:
        type: string
    type: object
  models.GeoPoint:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  users.GetProfileResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role:
        description: Should be 'customer', 'provider', or 'admin'
        type: string
    type: object
  users.UpdateProfileRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role:
        description: Should be 'customer', 'provider', or 'admin'
        type: string
      user_id:
        type: string
    type: object
  users.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role:
        description: Should be 'customer', 'provider', or 'admin'
        type: string
    type: object
...
  users.UserResponses:
    properties:
      users:
        items:
          $ref: '#/definitions/users.UserResponse'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gateway of On-Demand Car Wash Service
  title: On-Demand Car Wash Service
  version: "1.0"
paths:
  /admin/bookings/all/:id:
    get:
      description: Retrieves a list of bookings with pagination
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Page limit
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListBookingsResponse'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Lists bookings with pagination
      tags:
      - bookings
...
  /admin/user:
    put:
      description: Updates user info by ID
      parameters:
      - description: New user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.UpdateProfileRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Invalid user id
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates user
      tags:
      - admin
  /admin/user/{id}:
    delete:
      description: Deletes user info by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: User deleted successfully
          schema:
            type: string
        "400":
          description: Invalid user id
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Deletes user
      tags:
      - admin
    get:
      description: Retrieves user info by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.GetProfileResponse'
        "400":
          description: Invalid user id
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets user
      tags:
      - admin
  /admin/users:
    get:
      description: Retrieves users from the database by filtering
      parameters:
      - description: Full name
        in: query
        name: full_name
        type: string
      - description: Role
        in: query
        name: role
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of users per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponses'
        "400":
          description: Invalid pagination parameters
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Fetches users
      tags:
      - admin
  /customer/bookings:
    post:
      description: Creates a new booking and returns the booking details
      parameters:
      - description: Booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/models.Booking'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingResponse'
        "400":
          description: Invalid booking data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Creates a new booking
      tags:
      - bookings
  /customer/bookings/{id}:
    delete:
      description: Cancels booking by booking ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingResponse'
        "400":
          description: Invalid booking ID
          schema:
            type: string
        "404":
          description: Booking not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Cancels a booking by ID
      tags:
      - bookings
    get:
      description: Retrieves booking details by booking ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingResponse'
        "400":
          description: Invalid booking ID
          schema:
            type: string
        "404":
          description: Booking not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieves a booking by ID
      tags:
      - bookings
    put:
      description: Updates booking details by booking ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/booking.UpdateBookingRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingResponse'
        "400":
          description: Invalid booking data or ID
          schema:
            type: string
        "404":
          description: Booking not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates a booking by ID
      tags:
      - bookings
  /customer/bookings/all:
    get:
      description: Retrieves a list of bookings with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Page limit
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListBookingsResponse'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Lists bookings with pagination
      tags:
      - bookings
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"